requirements.txt
requests>=2.28.0
mysql-connector-python>=8.0.0
tabulate>=0.9.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
